+++ volume/samp_app/samp_app.jl
@@ -1,13 +1,16 @@
 using Genie
 import Genie.Router: route
-Genie.config.run_as_server = true
 Genie.REPL.load_app()
+Genie.config.run_as_server = true
 
-using SearchLight
-SearchLight.Configuration.load_db_connection()
 
 route("/hello") do
   "Hello world!"
 end
+
+using Questions
+# add initial seed data
+# Questions.seed()
+print(SearchLight.Database.query("select * from questions;"))
 
 Genie.AppServer.startup(8000, "0.0.0.0")
