+++ volume/samp_app/test/unit/questions_test.jl
@@ -1,0 +1,5 @@
+include(joinpath("..", "..", ".slbootstrap.jl"))
+using Test
+
+### Your tests here
+@test 1 == 1
